requirements.txt
transformers
torch
torchvision
Pillow
tqdm
flask
flask-sqlalchemy
flask-mail
flask-migrate
wtforms
werkzeug
numpy
pymysql
cryptography
email-validator
